+++ Kernel/headers/ports.h
@@ -8,11 +8,11 @@ uint8 read8bitport(uint16 portnumber);
 void write8bitportSlow(uint16 portnumber,uint8 data);
 
 void write16bitport(uint16 portnumber,uint16 data);
-uint8 read16bitport(uint16 portnumber);
+uint16 read16bitport(uint16 portnumber);
 void write16bitportSlow(uint16 portnumber,uint16 data);
 
 void write32bitport(uint16 portnumber,uint32 data);
-uint8 read32bitport(uint16 portnumber);
+uint32 read32bitport(uint16 portnumber);
 void write32bitportSlow(uint16 portnumber,uint32 data);
 
 #endif
